% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoData.R
\name{CompoData}
\alias{CompoData}
\title{CompoData}
\usage{
CompoData()
}
\description{
The main message of this package
}
\details{
The main message of this package is that it is simply a way to distribute a lecture
}
\author{
Emanuel Heitlinger
}
